Feat: Map Logic
